Turn applyRules into a processor
